Fix: crash when the input length was less than the terminal height
Fixed a crash that happened if we had content with fewer lines that fit in one page. We were trying to slice a vector from start to end where the end was always at least the terminal height. This was a problem when the content was actually less than the terminal's height.